fix(InputView) : mvc패턴에 따른 outputView 클래스 생성
