+++ src/main/java/RacingCarGameMain.java
@@ -2,8 +2,7 @@ import racingCar.Car;
 import racingCar.Cars;
 import utils.RandomUtils;
 import view.InputView;
-
-import java.util.List;
+import view.OutputView;
 
 public class RacingCarGameMain {
     private static final int MINIMUM_VALUE = 0;
@@ -16,54 +15,21 @@ public class RacingCarGameMain {
             Cars.carAdd(new Car(carName));
         }
         playGame(InputView.getTryNumber());
-        showResult();
+        OutputView.showResult();
     }
 
     private static void playGame(int count) {
         for (int i = 0; i < count; i++) {
-            playCar();
+            moveCar();
         }
     }
 
-    private static void playCar() {
+    private static void moveCar() {
         for (int i = 0; i < Cars.getSize(); i++) {
             Cars.getCars()
                     .get(i)
                     .move(RandomUtils.nextInt(MINIMUM_VALUE, MAXIMUM_VALUE));
         }
-        showStatus();
-    }
-
-    private static void showStatus() {
-        for (Car car : Cars.getCars()) {
-            System.out.println(car.getName() + " : " + bar(car.getDistance()));
-        }
-        System.out.println();
-    }
-
-    private static String bar(int distance) {
-        StringBuilder sb = new StringBuilder();
-        for (int i = 0; i < distance; i++) {
-            sb.append("-");
-        }
-        return sb.toString();
-    }
-
-    private static void showResult() {
-        List<String> winners = getWinners(findMax());
-        System.out.println(String.join(",", winners) + "가 최종 우승했습니다.");
-    }
-
-    private static List<String> getWinners(int max) {
-        List<String> winners = Cars.findWinners(max);
-        return winners;
-    }
-
-    private static int findMax() {
-        int maxNumber = 0;
-        for (Car car : Cars.getCars()) {
-            maxNumber = Math.max(maxNumber, car.getDistance());
-        }
-        return maxNumber;
+        OutputView.showStatus();
     }
 }
